tests/wpt/web-platform-tests/fullscreen/api/element-ready-check-fullscreen-iframe-child-manual.html

<!DOCTYPE html>
<title>Element ready check for child of a fullscreen iframe</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<iframe><!-- script inserts div here --></iframe>
<script>
// Verify that an iframe can itself go fullscreen, and that this doesn't
// influence the iframe ancestor test of the element ready check.
async_test(function(t)
{
    var iframe = document.querySelector("iframe");
    document.onfullscreenchange = t.step_func(function()
    {
        assert_equals(document.fullscreenElement, iframe, "fullscreen element");
        var div = document.createElement("div");
        // This adds the div to the iframe element itself, not to the iframe's
        // contentDocument. It's done here because the HTML parser treats the
        // content of iframe as a text node.
        iframe.appendChild(div);
        trusted_request(div, iframe.contentDocument.body);
        document.onfullscreenchange = t.unreached_func("second fullscreenchange event");
        document.onfullscreenerror = t.step_func_done();
    });
    trusted_request(iframe);
});
</script>
